requirements.txt
attrs==19.1.0
flask==1.0.2
freezegun==0.3.15
irc==17.0
mock==2.0.0
mypy-extensions==0.4.1
obs-websocket-py==0.5.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.20.0
websockets==8.1
